Fixed/Updated dice bug
Fixed dice to hide and display using the display:hidden and visable.
Cleaned up code by adding functions to reduce repeated code.

Todo:
- Add 
The program then provides the user with the options (i) to continue with another round, or (ii) to end the game. If option (i) is chosen, then the program remains in the play stage and proceeds with another round. If option (ii) is chosen, then the program proceeds to the end stage.

In the end stage, the program computes the average number of points won per round and displays

the number of rounds played,
the balance of points when reaching the end stage,
the average number of points won per round played, rounded to one digit after the decimal point.